Write more of the *.mp-data.
Oneway streets now work. DirIndicator is set, though I don't know
what it does. The road class from RouteParams is set in various
places, not really sure they should all get the same. (Why is it
"destination" class in an arc?) Speed is written. Forbidden
vehicle types are written, sorted out except for emergency and
delivery which appear to function differenlty.



git-svn-id: http://svn.mkgmap.org.uk/mkgmap/branches/nod@766 25d90789-57f7-4ee0-8453-03a3dfeeeb22

+++ src/uk/me/parabola/imgfmt/app/net/TableA.java
@@ -161,8 +161,8 @@ public class TableA {
 			log.debug("writing Table A entry", arcs.get(arc));
 			int pos = arc.roadDef.getOffsetNet1();
 			writer.put3(pos);
-			writer.put(arc.roadDef.getTabAInfo());
-			writer.put(arc.roadDef.getTabARestrictions());
+			writer.put((byte) arc.roadDef.getTabAInfo());
+			writer.put((byte) arc.roadDef.getTabARestrictions());
 		}
 	}
 }
